Added admin & image submission
added an admin option in AccountService so admin can choose that, type in an admin password (currently "00abc"), and then access the admin menu where they can approve images, view all submitted and approved images. And in HomeService, I added an option for users to submit an image. I'm using a scanner to take in any text input and store them in ArrayList, and then serialize them so the data can be seen again in the next program run.

Currently I'm facing these issues:
1. Serialized data get rewritten instead of added in after a new input is made in the next program run, this also happens when a user logout and login again to submit an image without exiting the program
2. An unchecked or unsafe operation warning in the methods I use to read and write the serialized data (storeImage.java)

I'm gonna see if i can find some ways to solve this but also thinking if it's necessary to solve it now since we have to change it to accept actual image files from users anyway (maybe by using FileChooser)